Add backend modifications after review
