+++ src/tasks/start/utils/backend/proxy.ts
@@ -1,7 +1,7 @@
 import { getMainContractName } from '../arapp'
 import { KernelInstance } from '~/typechain'
 import { logBack } from '../logger'
-import { getLog } from './helpers'
+import { getLog } from './logs'
 import Web3 from 'web3'
 import { TruffleEnvironmentArtifacts } from '@nomiclabs/buidler-truffle5/src/artifacts'
