+++ src/tasks/start/utils/backend/repo.ts
@@ -9,7 +9,7 @@ import {
 import Web3 from 'web3'
 import { TruffleEnvironmentArtifacts } from '@nomiclabs/buidler-truffle5/src/artifacts'
 import { logBack } from '../logger'
-import { getLog } from './helpers'
+import { getLog } from './logs'
 
 const ZERO_ADDR = '0x0000000000000000000000000000000000000000'
